Added Color primitive and moved DateTimeSpan primitive.
